Update and rename 01_CURE_ID_Cohort_20221130.sql to 01_CURE_ID_Cohort.sql
The insert into statement for the temporary table Vis_Occ threw an error. The script now explicitly mentions the columns of interest. 

The name is being changed to remove the date as changes will be made to this script as they arise. Please see the History to track the chronology of updates.