Update 07_A_condition_profile.sql
The prior version of the first block of this script counted only persons with exact matches to condition_concept_id's. For conditions recorded that were recorded in a fashion less specific than common practice, these conditions appeared to be uncommon. For example, tonsillitis may have been counted as a rare condition if most of the time patients were categorized as acute tonsillitis. This revision includes descendant concepts as well.